Split push_weights and reweight function
