+++ src/algorithms/mod.rs
@@ -11,6 +11,7 @@ mod isomorphic;
 mod projection;
 mod relabel_pairs;
 mod reverse;
+mod reweight;
 mod single_source_shortest_distance;
 mod union;
 mod weight_pushing;
@@ -28,6 +29,7 @@ pub use self::isomorphic::isomorphic;
 pub use self::projection::{project, project_input, project_output};
 pub use self::relabel_pairs::relabel_pairs;
 pub use self::reverse::reverse;
+pub use self::reweight::reweight;
 pub use self::single_source_shortest_distance::{
     shortest_distance, single_source_shortest_distance,
 };
