+++ src/algorithms/reweight.rs
@@ -1,0 +1,62 @@
+use crate::fst_traits::{ExpandedFst, FinalStatesIterator, Fst, MutableFst};
+use crate::semirings::{Semiring, WeaklyDivisibleSemiring};
+use crate::Result;
+
+macro_rules! state_to_dist {
+    ($state: expr, $dist: expr) => {
+        $dist
+            .get($state)
+            .ok_or_else(|| format_err!("State {} not in dists array", $state))?;
+    };
+}
+
+/// Reweights an FST according to a vector of potentials in a given direction.
+/// The weight must be left distributive when reweighting towards the initial
+/// state and right distributive when reweighting towards the final states.
+///
+/// An arc of weight w, with an origin state of potential p and destination state
+/// of potential q, is reweighted by p^-1 \otimes (w \otimes q) when reweighting
+/// torwards the initial state, and by (p \otimes w) \otimes q^-1 when
+/// reweighting towards the final states.
+pub fn reweight<F>(fst: &mut F, potentials: &Vec<F::W>) -> Result<()>
+where
+    F: Fst + ExpandedFst + MutableFst,
+    F::W: WeaklyDivisibleSemiring,
+{
+    let num_states = fst.num_states();
+
+    for state in 0..num_states {
+        let d_s = state_to_dist!(state, potentials);
+
+        if d_s.is_zero() {
+            continue;
+        }
+        for arc in fst.arcs_iter_mut(state)? {
+            let d_ns = state_to_dist!(arc.nextstate, potentials);
+            if d_ns.is_zero() {
+                continue;
+            }
+            arc.weight = d_s.inverse().times(&arc.weight.times(d_ns));
+        }
+    }
+
+    let final_states: Vec<_> = fst.final_states_iter().collect();
+
+    for final_state in final_states {
+        let d_s = state_to_dist!(final_state.state_id, potentials);
+        if d_s.is_zero() {
+            continue;
+        }
+        let new_weight = d_s.inverse().times(&final_state.final_weight);
+        fst.set_final(final_state.state_id, new_weight)?;
+    }
+
+    if let Some(start_state) = fst.start() {
+        let d_s = state_to_dist!(start_state, potentials);
+        for arc in fst.arcs_iter_mut(start_state)? {
+            arc.weight = arc.weight.times(d_s);
+        }
+    }
+
+    Ok(())
+}
